set static buildCrossChainBridgeTransaction method
